split constants files to match actions/reducers
