+++ react/constants/channel_action_types.js
@@ -1,0 +1,1 @@
+export const CHANNEL_UPDATE = 'CHANNEL_UPDATE';
