+++ react/constants/publish_claim_states.js
